+++ timeago_test.go
@@ -113,3 +113,33 @@ func TestTakeFunctionCanExceptTimestamp(t *testing.T) {
 	}
 
 }
+
+func TestTakeFunctionCanExceptTimePackage(t *testing.T) {
+	cases := []struct {
+		time   time.Time
+		result string
+	}{
+		{time.Now().Add(-time.Minute), "1 minute ago"},
+		{time.Now().Add(-time.Minute * 2), "2 minutes ago"},
+		{time.Now().Add(-time.Minute * 3), "3 minutes ago"},
+		{time.Now().Add(-time.Minute * 4), "4 minutes ago"},
+		{time.Now().Add(-time.Minute * 5), "5 minutes ago"},
+		{time.Now().Add(-time.Minute * 6), "6 minutes ago"},
+		{time.Now().Add(-time.Hour * 7), "7 hours ago"},
+		{time.Now().Add(-time.Hour * 8), "8 hours ago"},
+		{time.Now().Add(-time.Hour * 9), "9 hours ago"},
+		{time.Now().Add(-time.Hour * 10), "10 hours ago"},
+		{time.Now().Add(-time.Hour * 11), "11 hours ago"},
+	}
+
+	Set("language", "en")
+
+	for _, tc := range cases {
+		t.Run(tc.result, func(test *testing.T) {
+			if res := Take(tc.time); res != tc.result {
+				test.Errorf("Result must be %v, but got %v instead", tc.result, res)
+			}
+		})
+	}
+
+}